+++ internal/policies/mount/adsys_mount/src/logger.rs
@@ -1,24 +1,26 @@
 use chrono::Utc;
 
-use log::{Level, Log};
+use log::{Level, Log, Metadata, Record};
 /// Logger used by adsys_mount to provide context about the mount process.
 #[derive(Debug)]
 pub struct Logger {}
 
 impl Log for Logger {
-    fn enabled(&self, metadata: &log::Metadata) -> bool {
+    fn enabled(&self, metadata: &Metadata) -> bool {
         metadata.level() <= Level::Trace
     }
 
-    fn log(&self, record: &log::Record) {
-        if self.enabled(record.metadata()) {
-            eprintln!(
-                "{} - {}: {}",
-                Utc::now().format("%d/%m/%Y %H:%M:%S"),
-                record.level(),
-                record.args()
-            );
+    fn log(&self, record: &Record) {
+        if !self.enabled(record.metadata()) {
+            return;
         }
+
+        eprintln!(
+            "{} - {}: {}",
+            Utc::now().format("%d/%m/%Y %H:%M:%S"),
+            record.level(),
+            record.args()
+        );
     }
 
     fn flush(&self) {}
